Moved loadAbiFromFile to helper
