+++ package.json
@@ -36,6 +36,7 @@
     "esm": "^3.2.25",
     "husky": "^7.0.4",
     "js-yaml": "^4.1.0",
+    "keccak256": "^1.0.6",
     "mustache": "^4.2.0",
     "nodemon": "^2.0.18",
     "prettier": "^2.7.1",
